+++ modules/tinymce/src/themes/silver/main/ts/ui/header/InlineHeader.ts
@@ -1,6 +1,6 @@
 import { AlloyComponent, Boxes, Channels, Docking, VerticalDir } from '@ephox/alloy';
 import { Arr, Cell, Fun, Optional, Singleton } from '@ephox/katamari';
-import { Attribute, Css, Height, SugarBody, SugarElement, SugarLocation, Traverse, Width } from '@ephox/sugar';
+import { Attribute, Css, Height, Scroll, SugarBody, SugarElement, SugarLocation, Traverse, Width } from '@ephox/sugar';
 
 import DOMUtils from 'tinymce/core/api/dom/DOMUtils';
 import Editor from 'tinymce/core/api/Editor';
@@ -114,7 +114,7 @@ export const InlineHeader = (
     });
   };
 
-  const updateChromePosition = () => {
+  const updateChromePosition = (optToolbarWidth: Optional<number>) => {
     floatContainer.on((container) => {
       const toolbar = OuterContainer.getToolbar(mainUi.outerContainer);
       const offset = calcToolbarOffset(toolbar);
@@ -126,10 +126,46 @@ export const InlineHeader = (
         Math.max(targetBounds.y - Height.get(container.element) + offset, 0) :
         targetBounds.bottom;
 
-      Css.setAll(mainUi.outerContainer.element, {
+      const baseProperties = {
         position: 'absolute',
-        top: Math.round(top) + 'px',
-        left: Math.round(targetBounds.x) + 'px'
+        left: Math.round(targetBounds.x) + 'px',
+        top: Math.round(top) + 'px'
+      };
+
+      const widthProperties = optToolbarWidth.map(
+        (toolbarWidth: number) => {
+          const scroll = Scroll.get();
+
+          /*
+          As the editor container can wrap its elements (due to flex-wrap), the width of the container impacts also its height. Adding a minimum width works around two problems:
+
+          a) The docking behaviour (e.g. lazyContext) does not handle the situation of a very thin component near the edge of the screen very well, and actually has no concept of horizontal scroll - it only checks y values.
+
+          b) A very small toolbar is essentially unusable. On scrolling of X, we keep updating the width of the toolbar so that it can grow to fit the available space.
+
+          Note: this is entirely determined on the number of items in the menu and the toolbar, because when they wrap, that's what causes the height. Also, having multiple toolbars can also make it higher.
+          */
+          const minimumToolbarWidth = 150; // Value is arbitrary.
+
+          const availableWidth = window.innerWidth - (targetBounds.x - scroll.left);
+
+          const width = Math.max(
+            Math.min(
+              toolbarWidth,
+              availableWidth
+            ),
+            minimumToolbarWidth
+          );
+
+          return {
+            width: width + 'px'
+          };
+        }
+      ).getOr({ });
+
+      Css.setAll(mainUi.outerContainer.element, {
+        ...baseProperties,
+        ...widthProperties
       });
     });
   };
@@ -138,6 +174,22 @@ export const InlineHeader = (
     Arr.each(uiMotherships, (m) => {
       m.broadcastOn([ Channels.repositionPopups() ], { });
     });
+  };
+
+  const restoreAndGetCompleteOuterContainerWidth = (): Optional<number> => {
+    /*
+    Editors can be placed so far to the right that their left position is beyond the window width. This causes problems with flex-wrap. To solve this, set a width style on the container.
+    Natural width of the container needs to be calculated first.
+    */
+    if (!useFixedToolbarContainer) {
+      Css.set(mainUi.outerContainer.element, 'position', 'absolute');
+      Css.set(mainUi.outerContainer.element, 'left', '0px');
+      Css.remove(mainUi.outerContainer.element, 'width');
+      const w = Width.getOuter(mainUi.outerContainer.element);
+      return Optional.some(w);
+    } else {
+      return Optional.none();
+    }
   };
 
   const updateChromeUi = (resetDocking: boolean = false) => {
@@ -157,14 +209,20 @@ export const InlineHeader = (
       updateChromeWidth();
     }
 
-    // Refresh split toolbar
+    // This width can be used for calculating the "width" when resolving issues with flex-wrapping being triggered at the window width, despite scroll space being available to the right.
+    const optToolbarWidth: Optional<number> = restoreAndGetCompleteOuterContainerWidth();
+
+    /*
+    Refresh split toolbar. Before calling refresh, we need to make sure that we have the full width (through restoreAndGet.. above), otherwise too much will be put in the overflow drawer.
+    */
     if (isSplitToolbar) {
       OuterContainer.refreshToolbar(mainUi.outerContainer);
     }
 
     // Positioning
     if (!useFixedToolbarContainer) {
-      updateChromePosition();
+      // This will position the container in the right spot.
+      updateChromePosition(optToolbarWidth);
     }
 
     // Docking